On peut ajouter des objets à une page !

+++ database/migrations/2019_11_23_200627_create_inventories_table.php
@@ -14,10 +14,10 @@ class CreateInventoriesTable extends Migration
     public function up()
     {
         Schema::create('inventories', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->unsignedBigInteger('character_id');
             $table->foreign('character_id')->references('id')->on('characters');
-            $table->unsignedBigInteger('item_id');
+            $table->unsignedInteger('item_id');
             $table->foreign('item_id')->references('id')->on('items');
             $table->integer('quantity')->default(0);
             $table->boolean('used')->default(false);
